safari-534.53-branch/LayoutTests/fast/dom/Geolocation/maximum-age-expected.txt

Tests that the PositionOptions.maximumAge parameter is correctly applied.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy

PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy

PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy

PASS error.code is mockCode
PASS error.message is mockMessage
PASS successfullyParsed is true

TEST COMPLETE
